introduced BaseSpan
