+++ CHANGELOG.asciidoc
@@ -25,7 +25,7 @@ endif::[]
 
 [float]
 ===== Breaking changes
-* The following public API types were `public` so far and became package-private: `NoopScope`, `ScopeImpl` and `AbstractSpanImpl`.
+* The following public API types were `public` so far and became package-private: `NoopScope`, `ScopeImpl` and `BaseAbstractSpanImpl`.
   If your code is using them, you will need to change that when upgrading to this version.
   Related PR: {pull}1532[#1532]
