+++ apm-agent-plugins/apm-api-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -7,32 +7,32 @@ co.elastic.apm.agent.pluginapi.TransactionInstrumentation$SetUserInstrumentation
 co.elastic.apm.agent.pluginapi.TransactionInstrumentation$EnsureParentIdInstrumentation
 co.elastic.apm.agent.pluginapi.TransactionInstrumentation$SetResultInstrumentation
 co.elastic.apm.agent.pluginapi.TransactionInstrumentation$AddCustomContextInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$SetNameInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$SetTypeInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$SetTypesInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$DoCreateSpanInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$SetStartTimestampInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$EndInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$EndWithTimestampInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$CaptureExceptionInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$GetIdInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$GetTraceIdInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$AddStringLabelInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$AddNumberLabelInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$AddBooleanLabelInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$ActivateInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$IsSampledInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$InjectTraceHeadersInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$WithDestinationServiceNameInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$WithDestinationServiceResourceInstrumentation
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$WithDestinationServiceTypeInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$SetNameInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$SetTypeInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$SetTypesInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$DoCreateSpanInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$SetStartTimestampInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$EndInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$EndWithTimestampInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$CaptureExceptionInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$GetIdInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$GetTraceIdInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$AddStringLabelInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$AddNumberLabelInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$AddBooleanLabelInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$ActivateInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$IsSampledInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$InjectTraceHeadersInstrumentation
+co.elastic.apm.agent.pluginapi.SpanInstrumentation$WithDestinationServiceNameInstrumentation
+co.elastic.apm.agent.pluginapi.SpanInstrumentation$WithDestinationServiceResourceInstrumentation
+co.elastic.apm.agent.pluginapi.SpanInstrumentation$WithDestinationServiceTypeInstrumentation
 co.elastic.apm.agent.pluginapi.CaptureExceptionInstrumentation
 co.elastic.apm.agent.pluginapi.ApiScopeInstrumentation
 co.elastic.apm.agent.pluginapi.CaptureTransactionInstrumentation
 co.elastic.apm.agent.pluginapi.CaptureSpanInstrumentation
 co.elastic.apm.agent.pluginapi.TracedInstrumentation
 # legacy
-co.elastic.apm.agent.pluginapi.AbstractSpanInstrumentation$LegacyCaptureExceptionInstrumentation
+co.elastic.apm.agent.pluginapi.BaseAbstractSpanInstrumentation$LegacyCaptureExceptionInstrumentation
 co.elastic.apm.agent.pluginapi.LegacySpanInstrumentation$SetNameInstrumentation
 co.elastic.apm.agent.pluginapi.LegacySpanInstrumentation$SetTypeInstrumentation
 co.elastic.apm.agent.pluginapi.LegacySpanInstrumentation$DoCreateSpanInstrumentation
